requirements.txt
python-telegram-bot
aiohttp
aiosqlite
nest-asyncio
